First pass at moving StaticArray to core
