+++ src/axom/core/CMakeLists.txt
@@ -63,6 +63,7 @@ set(core_headers
     Path.hpp
     RangeAdapter.hpp
     StackArray.hpp
+    StaticArray.hpp
     Types.hpp
     memory_management.hpp
